+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+0.6.2.1
+---
+* Fixed error output for unexpected floating point literal
+
 0.6.2
 ---
 * Nicer errors on unterminated inline lists and sections.